+++ quilt-sync/CHANGELOG.md
@@ -9,6 +9,12 @@
 <!-- markdownlint-disable MD013 -->
 # Changelog
 
+## [v0.17.2-alpha2] - 2026-05-13
+
+### Added
+
+- Filesystem watcher refreshes local package status on disk changes (default on; toggle in Settings)
+
 ## [v0.17.2-alpha1] - 2026-05-13
 
 ### Added